Move configure settings logic to settings module
