+++ todo_project/settings/configure.py
